requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10 --only-binary :all:
python-dotenv==1.0.1
openai==1.26.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.125.0
python-dateutil==2.9.0.post0
dateparser==1.2.0
